move family detail loading to shared
